Create Contact and extract commonalities from Address,Email,Phone
